Switch from log4j to console logging for Spark.
This is required because on CDH5, using Log4j with Spark+YARN will
create conflicts with the logging configuration of the YARN container,
causing fatal problems when trying to start up a Shark application
master due to permissions and file conflicts.

Use SLF4J-Simple instead of SLF4J-Log4j to accomplish this.